# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MixPlay API",
        "description": "API da aplicação MixPlay, para pegar os dados.",
        "contact": {
            "email": "rodrigo.pinheiroa@hotmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Default Server"
        }
    ],
    "paths": {
...
        "/movie": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Listar filmes",
                "operationId": "movieIndex",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Movies"
                ],
                "summary": "Cadastrar um filme",
                "operationId": "movieStore",
                "requestBody": {
                    "$ref": "#components/requestBodies/StoreMovie"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            }
        },
        "/movie/{movie}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Exibir detalhes de um filme",
                "operationId": "movieShow",
                "parameters": [
                    {
                        "name": "movie",
                        "in": "path",
                        "description": "ID do filme",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "Movies"
                ],
                "summary": "Atualizar filme",
                "operationId": "movieUpdate",
                "parameters": [
                    {
                        "name": "movie",
                        "in": "path",
                        "description": "ID do filme",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#components/requestBodies/StoreMovie"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movies"
                ],
                "summary": "Remover filme",
                "operationId": "movieDestroy",
                "parameters": [
                    {
                        "name": "movie",
                        "in": "path",
                        "description": "ID do filme",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Movie": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}